renamed tool router and controller and fixed syntax error
