+++ src/routes/bmdashboard/bmToolRouter.js
@@ -2,7 +2,7 @@ const express = require('express');
 
 const routes = function (BuildingTool) {
     const toolRouter = express.Router();
-    const controller = require('../../controllers/bmdashboard/bmToolsController')(BuildingTool);
+    const controller = require('../../controllers/bmdashboard/bmToolController')(BuildingTool);
 
     toolRouter.route('/tools/:toolId')
         .get(controller.fetchSingleTool);
